Address PR review: preserve chart range on timezone toggle.
Refresh only the tick formatter when switching UTC/Local, and move timestamp tests into time.test.ts with an exact America/New_York local assertion.
